third_party/WebKit/LayoutTests/compositing/overflow/scrolls-with-respect-to-expected.txt

This test ensures that Layer::scrollsWithRespectTo always returns the correct answer on relatively "basic" elements as well as fixed- and absolute-positioned elements (ie, those with non-obvious containing blocks), and iframes.

predecessor does not scroll with respect to predecessor
predecessor scrolls with respect to container
predecessor scrolls with respect to scrolled1
predecessor scrolls with respect to scrolled2
predecessor scrolls with respect to scrolled3
predecessor does not scroll with respect to fixed-descendant
predecessor scrolls with respect to successor
predecessor scrolls with respect to root
predecessor scrolls with respect to descendant
container scrolls with respect to predecessor
container does not scroll with respect to container
container scrolls with respect to scrolled1
container scrolls with respect to scrolled2
container scrolls with respect to scrolled3
container scrolls with respect to fixed-descendant
container does not scroll with respect to successor
container does not scroll with respect to root
container scrolls with respect to descendant
scrolled1 scrolls with respect to predecessor
scrolled1 scrolls with respect to container
scrolled1 does not scroll with respect to scrolled1
scrolled1 does not scroll with respect to scrolled2
scrolled1 does not scroll with respect to scrolled3
scrolled1 scrolls with respect to fixed-descendant
scrolled1 scrolls with respect to successor
scrolled1 scrolls with respect to root
scrolled1 does not scroll with respect to descendant
scrolled2 scrolls with respect to predecessor
scrolled2 scrolls with respect to container
scrolled2 does not scroll with respect to scrolled1
scrolled2 does not scroll with respect to scrolled2
scrolled2 does not scroll with respect to scrolled3
scrolled2 scrolls with respect to fixed-descendant
scrolled2 scrolls with respect to successor
scrolled2 scrolls with respect to root
scrolled2 does not scroll with respect to descendant
scrolled3 scrolls with respect to predecessor
scrolled3 scrolls with respect to container
scrolled3 does not scroll with respect to scrolled1
scrolled3 does not scroll with respect to scrolled2
scrolled3 does not scroll with respect to scrolled3
scrolled3 scrolls with respect to fixed-descendant
scrolled3 scrolls with respect to successor
scrolled3 scrolls with respect to root
scrolled3 does not scroll with respect to descendant
fixed-descendant does not scroll with respect to predecessor
fixed-descendant scrolls with respect to container
fixed-descendant scrolls with respect to scrolled1
fixed-descendant scrolls with respect to scrolled2
fixed-descendant scrolls with respect to scrolled3
fixed-descendant does not scroll with respect to fixed-descendant
fixed-descendant scrolls with respect to successor
fixed-descendant scrolls with respect to root
fixed-descendant scrolls with respect to descendant
successor scrolls with respect to predecessor
successor does not scroll with respect to container
successor scrolls with respect to scrolled1
successor scrolls with respect to scrolled2
successor scrolls with respect to scrolled3
successor scrolls with respect to fixed-descendant
successor does not scroll with respect to successor
successor does not scroll with respect to root
successor scrolls with respect to descendant
root scrolls with respect to predecessor
root does not scroll with respect to container
root scrolls with respect to scrolled1
root scrolls with respect to scrolled2
root scrolls with respect to scrolled3
root scrolls with respect to fixed-descendant
root does not scroll with respect to successor
root does not scroll with respect to root
root scrolls with respect to descendant
descendant scrolls with respect to predecessor
descendant scrolls with respect to container
descendant does not scroll with respect to scrolled1
descendant does not scroll with respect to scrolled2
descendant does not scroll with respect to scrolled3
descendant scrolls with respect to fixed-descendant
descendant scrolls with respect to successor
descendant scrolls with respect to root
descendant does not scroll with respect to descendant
iframe-div does not scroll with respect to iframe-div
iframe-div does not scroll with respect to iframe-abs-div
iframe-div scrolls with respect to iframe-fixed-div
iframe-div does not scroll with respect to iframe-root
iframe-abs-div does not scroll with respect to iframe-div
iframe-abs-div does not scroll with respect to iframe-abs-div
iframe-abs-div scrolls with respect to iframe-fixed-div
iframe-abs-div does not scroll with respect to iframe-root
iframe-fixed-div scrolls with respect to iframe-div
iframe-fixed-div scrolls with respect to iframe-abs-div
iframe-fixed-div does not scroll with respect to iframe-fixed-div
iframe-fixed-div scrolls with respect to iframe-root
iframe-root does not scroll with respect to iframe-div
iframe-root does not scroll with respect to iframe-abs-div
iframe-root scrolls with respect to iframe-fixed-div
iframe-root does not scroll with respect to iframe-root
